requirements.txt
# requirements.txt
# Note: To avoid Pillow build errors on Python 3.13, install facenet-pytorch separately:
#       python -m pip install facenet-pytorch --no-deps

torch>=2.0.0
torchvision>=0.15.0
pytorch-lightning>=2.0.0
timm>=0.6.0
scikit-learn>=1.3.0
numpy>=1.24.0,<2.0.0
Pillow==10.1.1
# facenet-pytorch is installed separately to prevent dependency conflicts
